[BUG FIX] #41: dev merge 충돌로 인한 테스트 미통과 해결
