+++ backend/src/main/java/autoever2/cartag/repository/OptionRepository.java
@@ -161,5 +161,4 @@ public class OptionRepository {
 
         return template.queryForObject(sql, param, Long.class);
     }
-
 }
