+++ backend/src/test/java/autoever2/cartag/controller/CarControllerTest.java
@@ -4,11 +4,10 @@ import autoever2.cartag.domain.car.CarDto;
 import autoever2.cartag.domain.car.CarTypeDto;
 import autoever2.cartag.domain.car.TrimDefaultOptionDto;
 import autoever2.cartag.domain.option.QuoteSubOptionDto;
-import autoever2.cartag.domain.share.QuoteIdList;
-import autoever2.cartag.domain.share.QuoteInfoDto;
+import autoever2.cartag.domain.quote.QuoteDataDto;
+import autoever2.cartag.domain.quote.QuoteInfoDto;
 import autoever2.cartag.service.CarService;
-import org.json.JSONArray;
-import org.json.JSONObject;
+import autoever2.cartag.service.QuoteService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -20,10 +19,8 @@ import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.ResultActions;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 
-import java.awt.image.AreaAveragingScaleFilter;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Stack;
 
 import static org.mockito.BDDMockito.given;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
@@ -31,6 +28,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 
 @WebMvcTest(CarController.class)
 class CarControllerTest {
+
     @Autowired
     MockMvc mockMvc;
 
@@ -39,8 +37,6 @@ class CarControllerTest {
 
     private List<CarDto> carDtoList;
     private List<TrimDefaultOptionDto> defaultOptions;
-    private List<QuoteSubOptionDto> quoteSubOptionDtoList;
-    private QuoteInfoDto quoteInfoDto;
 
     @BeforeEach
     void setup() {
@@ -112,70 +108,6 @@ class CarControllerTest {
                 .build()
         );
 
-        quoteSubOptionDtoList = new ArrayList<>();
-
-        quoteSubOptionDtoList.add(QuoteSubOptionDto
-                .builder()
-                .optionId(69)
-                .optionPrice(123400L)
-                .optionTitle("상세품목")
-                .optionImage("image_1.jpg")
-                .optionName("3열 열선시트")
-                .build());
-
-        quoteSubOptionDtoList.add(QuoteSubOptionDto
-                .builder()
-                .optionId(70)
-                .optionPrice(100L)
-                .optionTitle("휠")
-                .optionImage("iamge_2.jpg")
-                .optionName("20인치 다크 스퍼터링")
-                .build());
-
-        quoteSubOptionDtoList.add(QuoteSubOptionDto
-                .builder()
-                .optionId(71)
-                .optionPrice(8977L)
-                .optionTitle("악세사리")
-                .optionImage("image_3.jpg")
-                .optionName("후진가이드램프")
-                .build());
-
-        quoteInfoDto = QuoteInfoDto
-                .builder()
-                .carId(1)
-                .trim("Le Blanc")
-                .carDefaultPrice(40000000)
-                .powerTrainId(1)
-                .powerTrainTitle("파워트레인")
-                .powerTrainImage("image_1.jpg")
-                .powerTrainName("디젤2.2")
-                .powerTrainPrice(0L)
-                .bodyTypeId(5)
-                .bodyTypeTitle("바디타입")
-                .bodyTypeImage("image_2.jpg")
-                .bodyTypeName("7인승")
-                .bodyTypePrice(1500L)
-                .operationTitle("구동방식")
-                .operationId(3)
-                .operationImage("image_3.jpg")
-                .operationName("2WD")
-                .operationPrice(8900L)
-                .colorOuterTitle("외장색상")
-                .colorOuterId(4)
-                .colorOuterImage("red_1.jpg")
-                .colorOuterPrice(1500L)
-                .colorOuterImageName("퍼플 펄")
-                .colorCarOuterImage("outer_red.jpg")
-                .colorInnerTitle("내장색상")
-                .colorInnerId(1)
-                .colorInnerImage("blue_1.jpg")
-                .colorInnerPrice(1220L)
-                .colorInnerImageName("퀄팅 천연(파랑)")
-                .colorCarInnerImage("inner_red.jpg")
-                .optionList(quoteSubOptionDtoList)
-                .build();
-
     }
 
     @Test
@@ -224,46 +156,5 @@ class CarControllerTest {
                 .andExpect(jsonPath("$[0].carTypeId").value(1))
                 .andExpect(jsonPath("$[1].carTypeImage").value("/cartype/santafe.png"))
                 .andExpect(jsonPath("$[2].carTypeName").value("디 올 뉴 코나 Hybrid"));
-    }
-
-    @Test
-    @DisplayName("id에 따른 공유 데이터 반환")
-    void getShareInfo() throws Exception {
-        ArrayList<Integer> optionIds = new ArrayList<>();
-        optionIds.add(69);
-        optionIds.add(70);
-        optionIds.add(71);
-        QuoteIdList quoteIdList = QuoteIdList
-                .builder()
-                .carId(1)
-                .powerTrainId(1)
-                .operationId(3)
-                .bodyTypeId(5)
-                .innerColorId(1)
-                .outerColorId(4)
-                .optionIdList(optionIds)
-                .build();
-
-        given(service.findShareInfoDto(quoteIdList)).willReturn(quoteInfoDto);
-
-        ResultActions resultActions = mockMvc.perform(MockMvcRequestBuilders.post("/api/cars/infos/shares")
-                .content("{\n" +
-                        "    \"carId\": 1,\n" +
-                        "    \"powerTrainId\": 1,\n" +
-                        "    \"bodyTypeId\": 5,\n" +
-                        "    \"operationId\": 3,\n" +
-                        "    \"outerColorId\": 4,\n" +
-                        "    \"innerColorId\": 1,\n" +
-                        "    \"optionIdList\": [69, 70, 71]\n" +
-                        "}").contentType(MediaType.APPLICATION_JSON));
-        resultActions.andExpect(status().isOk())
-                .andExpect(jsonPath("$.carDefaultPrice").value(40000000))
-                .andExpect(jsonPath("$.trim").value("Le Blanc"))
-                .andExpect(jsonPath("$.powerTrainImage").value("image_1.jpg"))
-                .andExpect(jsonPath("$.bodyTypeName").value("7인승"))
-                .andExpect(jsonPath("$.operationTitle").value("구동방식"))
-                .andExpect(jsonPath("$.colorOuterPrice").value(1500L))
-                .andExpect(jsonPath("$.colorCarInnerImage").value("inner_red.jpg"))
-                .andExpect(jsonPath("$.optionList.size()").value(3));
     }
 }
