+++ backend/src/test/java/autoever2/cartag/controller/ModelControllerTest.java
@@ -3,9 +3,7 @@ package autoever2.cartag.controller;
 import autoever2.cartag.domain.model.ModelDetailMappedDto;
 import autoever2.cartag.domain.model.ModelEfficiencyDataDto;
 import autoever2.cartag.domain.model.ModelShortDataDto;
-import autoever2.cartag.domain.model.PowerTrainMappedDto;
 import autoever2.cartag.service.ModelService;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
